+++ lib/compiler/include/compiler/ast/statement.hh
@@ -27,8 +27,9 @@ struct block_stmt {
     MAKE_ITERATOR(statements_t, std::vector<stmt_handle>, statements);
 
     statements_t statements;
+    bool         is_constexpr{false};
 
-    [[nodiscard]] static auto parse(syntax::parser& parser)
+    [[nodiscard]] static auto parse(syntax::parser& parser, bool is_constexpr = false)
         -> stdx::result<stmt_handle, syntax::diagnostic>;
 };